+++ src/main/java/com/jovan/erp_v1/repository/StorageRepository.java
@@ -1,5 +1,6 @@
 package com.jovan.erp_v1.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -13,19 +14,20 @@ import com.jovan.erp_v1.model.Storage;
 @Repository
 public interface StorageRepository extends JpaRepository<Storage, Long> {
 
-	List<Storage> findByType(StorageType type);
-    
+    List<Storage> findByType(StorageType type);
+
     List<Storage> findByName(String name);
-    
+
     List<Storage> findByLocation(String location);
-    
-    List<Storage> findByCapacity(Double capacity);
-    
+
+    List<Storage> findByCapacity(BigDecimal capacity);
+
     List<Storage> findByNameAndLocation(String name, String location);
+
     List<Storage> findByTypeAndCapacityGreaterThan(StorageType type, Double capacity);
-    
+
     @Query("SELECT s FROM Storage s WHERE SIZE(s.goods) > :minCount")
     List<Storage> findStoragesWithMinGoods(@Param("minCount") int minCount);
-    
+
     List<Storage> findByNameContainingIgnoreCase(String name);
 }
