+++ src/main/java/com/jovan/erp_v1/request/SupplyRequest.java
@@ -1,5 +1,6 @@
 package com.jovan.erp_v1.request;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -8,19 +9,13 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 public record SupplyRequest(
-		
+
 		Long id,
-		@NotNull(message = "ID skladišta je obavezan")
-		Long storageId,
+		@NotNull(message = "ID skladišta je obavezan") Long storageId,
 
-		@NotEmpty(message = "Lista robe je obavezna")
-		List<Long> goodsIds,
+		@NotEmpty(message = "Lista robe je obavezna") List<Long> goodsIds,
 
-		@NotNull(message = "Količina je obavezna")
-		@Positive(message = "Količina mora biti veća od 0")
-		Integer quantity,
+		@NotNull(message = "Količina je obavezna") @Positive(message = "Količina mora biti veća od 0") BigDecimal quantity,
 
-		@NotNull(message = "Datum ažuriranja je obavezan")
-		LocalDateTime updates
-		) {
+		@NotNull(message = "Datum ažuriranja je obavezan") LocalDateTime updates) {
 }
